hw3: refactor поправил скелетон карточки
