+++ src/App/pages/ProductPage/components/RelatedItems/RelatedItems.module.scss
@@ -4,10 +4,11 @@
   &__list {
     display: flex;
     flex-direction: row;
-    gap: calc((100% - vars.$width-card * 3) / 2);
+    gap: calc((100% - vars.$card-width * 3) / 2);
     width: 100%;
     margin-top: 32px;
     padding: 0;
+    padding-bottom: 10px;
     overflow-x: scroll;
   }
 }
