+++ src/App/styles/variables.scss
@@ -40,7 +40,8 @@ $button-disabled-bg: $color-card-bg;
   margin: 0 auto;
 }
 
-$width-card: 360px;
+$card-width: 360px;
+$card-height: 622px;
 
 @mixin skeleton($width: 100%, $height: 100px) {
   width: $width;
